rename images.py
