+++ mimic/model/rackspace_images.py
